refactor: compartmentalized to allow query to refetch
