+++ src/example-app/package.json
@@ -12,7 +12,8 @@
     "next": "15.1.7",
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
-    "react-player": "^2.16.0"
+    "react-player": "^2.16.0",
+    "react-query": "^3.39.3"
   },
   "devDependencies": {
     "@eslint/eslintrc": "^3",
